chore(kitchensink): abstract pagination controls
chore(kitchensink): abstract LoadMore button

chore(kitchensink): use json document editor
