+++ apps/kitchensink-react/src/DocumentCollection/DocumentProjectionRoute.tsx
@@ -1,9 +1,11 @@
 import {DocumentHandle, useDocumentProjection, usePaginatedDocuments} from '@sanity/sdk-react'
-import {Box, Button, Card, Flex, Label, Spinner, Stack, Text, TextInput} from '@sanity/ui'
+import {Box, Button, Card, Flex, Spinner, Stack, Text} from '@sanity/ui'
 import groq, {defineProjection} from 'groq'
 import {JSX, ReactNode, Suspense, useRef, useState} from 'react'
 import {ErrorBoundary} from 'react-error-boundary'
 
+import {PaginatedListToolbar} from '../components/PaginatedListToolbar'
+import {PaginationControls} from '../components/PaginationControls'
 // Import the custom table components
 import {Table, TD, TH, TR} from '../components/TableElements'
 
@@ -133,111 +135,6 @@ function AuthorRow({
 }
 
 // Define interface for pagination controls props
-interface PaginationControlsProps {
-  currentPage: number
-  totalPages: number
-  hasFirstPage: boolean
-  hasPreviousPage: boolean
-  hasNextPage: boolean
-  hasLastPage: boolean
-  firstPage: () => void
-  previousPage: () => void
-  nextPage: () => void
-  lastPage: () => void
-  goToPage: (pageNumber: number) => void
-  isPending: boolean
-}
-
-// Pagination controls component
-function PaginationControls({
-  currentPage,
-  totalPages,
-  hasFirstPage,
-  hasPreviousPage,
-  hasNextPage,
-  hasLastPage,
-  firstPage,
-  previousPage,
-  nextPage,
-  lastPage,
-  goToPage,
-  isPending,
-}: PaginationControlsProps) {
-  const buttonStyle = {
-    minWidth: '40px',
-    margin: '0 4px',
-    textAlign: 'center',
-  } as const
-
-  // Generate page number buttons
-  const pageButtons = () => {
-    const buttons = []
-    const maxVisiblePages = 5
-    let startPage = Math.max(1, currentPage - Math.floor(maxVisiblePages / 2))
-    const endPage = Math.min(totalPages, startPage + maxVisiblePages - 1)
-
-    if (endPage - startPage + 1 < maxVisiblePages) {
-      startPage = Math.max(1, endPage - maxVisiblePages + 1)
-    }
-
-    for (let i = startPage; i <= endPage; i++) {
-      buttons.push(
-        <Button
-          key={i}
-          mode={i === currentPage ? 'default' : 'ghost'}
-          onClick={() => goToPage(i)}
-          style={{
-            ...buttonStyle,
-            fontWeight: i === currentPage ? 'bold' : 'normal',
-          }}
-        >
-          {i}
-        </Button>,
-      )
-    }
-    return buttons
-  }
-
-  return (
-    <Flex align="center" justify="space-between" padding={3}>
-      <Flex>
-        <Button
-          onClick={firstPage}
-          disabled={!hasFirstPage}
-          style={buttonStyle}
-          text="<<"
-          mode="ghost"
-        />
-        <Button
-          onClick={previousPage}
-          disabled={!hasPreviousPage}
-          style={buttonStyle}
-          text="<"
-          mode="ghost"
-        />
-        {pageButtons()}
-        <Button
-          onClick={nextPage}
-          disabled={!hasNextPage}
-          style={buttonStyle}
-          text=">"
-          mode="ghost"
-        />
-        <Button
-          onClick={lastPage}
-          disabled={!hasLastPage}
-          style={buttonStyle}
-          text=">>"
-          mode="ghost"
-        />
-      </Flex>
-      <Text size={1} style={{opacity: isPending ? 0.5 : 1}}>
-        Page {currentPage} of {totalPages}
-      </Text>
-    </Flex>
-  )
-}
-
 export function DocumentProjectionRoute(): JSX.Element {
   const [searchTerm, setSearchTerm] = useState('')
   const [pageSize, setPageSize] = useState(5)
@@ -282,46 +179,16 @@ export function DocumentProjectionRoute(): JSX.Element {
     <Stack space={4}>
       <Card padding={4}>
         <Stack space={4}>
-          <Flex justify="space-between" align="center">
-            <Box style={{width: '300px'}}>
-              <Label htmlFor="search" size={1} style={{marginBottom: '4px', display: 'block'}}>
-                Search Authors
-              </Label>
-              <TextInput
-                id="search"
-                value={searchTerm}
-                onChange={handleSearchChange}
-                placeholder="Search by name..."
-                style={{width: '100%'}}
-              />
-            </Box>
-            <Box>
-              <Label htmlFor="pageSize" size={1} style={{marginBottom: '4px', display: 'block'}}>
-                Items per page
-              </Label>
-              <select
-                id="pageSize"
-                value={pageSize}
-                onChange={handlePageSizeChange}
-                style={{
-                  padding: '8px',
-                  borderRadius: '4px',
-                  border: '1px solid #ccc',
-                }}
-              >
-                <option value={5}>5</option>
-                <option value={10}>10</option>
-                <option value={25}>25</option>
-                <option value={50}>50</option>
-              </select>
-            </Box>
-          </Flex>
-
-          <Box style={{borderRadius: '4px', border: '1px solid #eee', padding: '8px'}}>
-            <Text size={1}>
-              Showing {startIndex + 1}-{Math.min(endIndex, count)} of {count} authors
-            </Text>
-          </Box>
+          <PaginatedListToolbar
+            noun="authors"
+            searchTerm={searchTerm}
+            onSearchChange={handleSearchChange}
+            pageSize={pageSize}
+            onPageSizeChange={handlePageSizeChange}
+            count={count}
+            startIndex={startIndex}
+            endIndex={endIndex}
+          />
 
           <PaginationControls
             currentPage={currentPage}
